pyproject.toml
[project]
name = "pdf2texttranslator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.98.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
]
